This prevents webpack from trying to minify the yargs dependency when deploying typecript-to-schema as a library in a web app.
This prevents webpack from trying to minify the yargs dependency when deploying typecript-to-schema as a library in a web app.

yargs cannot be minified because it's published to npm as es6.
