+++ bin/typescript-json-schema
@@ -1,3 +1,3 @@
 #!/usr/bin/env node
-var app = require('../typescript-json-schema.js');
+var app = require('../typescript-json-schema-cli.js');
 app.run();
